feat: implement JWT auth, user context, and upgrade to .NET 9

+++ backend/PharmaCore.Infrastructure/Data/Configurations/UserConfiguration.cs
@@ -43,7 +43,7 @@ namespace PharmaCore.Infrastructure.Data.Configurations
 
             builder.Property(u => u.Role)
                    .IsRequired()
-                   .HasDefaultValue((byte)UserRole.Pharmacist);
+                   .HasDefaultValue(UserRole.Pharmacist);
             builder.HasIndex(u => u.Role);
 
             // Default Values for Flags
